Property change change
Edited the property change method on the chat client to send out a message json when a message is sent instead of two strings. I also used the fire that occurs when a username is set to make the server send out a broadcast with the new username when that occurs.